+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## TODO
+- real multithread
+- const cast
+- timeout
+- ICP as a Class
+- ICP multithread
+
+## v2.1.7 - 2024-12-26
+- [new feature] optional ICP algorithm for concatinate pointclouds
+
 ## v2.1.6 - 2024-12-25
 - [new feature] support for random num of pointcloud input sources
 - [bug fix] a single `tf::TransformListener` is sufficient
@@ -20,14 +30,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [upgrade] check if both `frame_id`s are the same
 - [upgrade] configurable `tf_duration`
 
-## v2.1.3: 2024-08-23
+## v2.1.3 - 2024-08-23
 - [new feature] 转发lidar_frame坐标系下的点云到world坐标系下
 
-## v2.1.2: 
+## v2.1.2
 - import `message_filters::sync_policies::ApproximateTime` for merge
 
-## v2.1.1: 
+## v2.1.1
 - publish pose
 
-## v2.1.0: 
-- merge `PointCloud2` from several sources, with `frame_id = "base_link"`
+## v2.1.0
+- merge 2 `PointCloud2`s, with fixed `frame_id = "base_link"`